requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
backports.zoneinfo==0.2.1; python_version < '3.9'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
boto3==1.34.88; python_version >= '3.8'
botocore==1.34.88; python_version >= '3.8'
django==4.2.11; python_version >= '3.8'
django-bootstrap4==24.3; python_version >= '3.8'
django-environ==0.11.2; python_version >= '3.6' and python_version < '4'
djangorestframework==3.15.1; python_version >= '3.6'
drf-yasg==1.21.7; python_version >= '3.6'
inflection==0.5.1; python_version >= '3.5'
jmespath==1.0.1; python_version >= '3.7'
packaging==24.0; python_version >= '3.7'
psycopg2-binary==2.9.9; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2024.1
pyyaml==6.0.1; python_version >= '3.6'
s3transfer==0.10.1; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.5; python_version >= '3.8'
sqlparse==0.5.0; python_version >= '3.8'
typing-extensions==4.11.0; python_version < '3.11'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.18; python_version < '3.10'
